32 vs 64 Bit warnings removal - take 2
git-svn-id: file:///home/jj/hercules.svn/trunk@2425 956126f8-22a0-4046-8f4a-272fa8102e63

+++ bldcfg.c
@@ -228,7 +228,7 @@ int off;
     }
 
     /* Trying to get mainstor aligned to the next 4K boundary - Greg */
-    off = (int)sysblk.mainstor & 0xFFF;
+    off = (long)sysblk.mainstor & 0xFFF;
     sysblk.mainstor += off ? 4096 - off : 0;
 
     /* Obtain main storage key array */
